Fix decimals test as its return type changed to BigInteger from int

+++ irc2-token/src/test/java/com/iconloop/score/example/IRC2BasicTest.java
@@ -58,7 +58,7 @@ class IRC2BasicTest extends TestBase {
 
     @Test
     void decimals() {
-        assertEquals(decimals, tokenScore.call("decimals"));
+        assertEquals(BigInteger.valueOf(decimals), tokenScore.call("decimals"));
     }
 
     @Test
